wii-esc-ng:
EEPROM support: Small refactoring

git-svn-id: https://wii-esc.googlecode.com/svn/branches/wii-esc-ng@312 5c42c5a8-900f-5b5a-2f8c-d720ba63385b

+++ src/global.h
@@ -51,8 +51,7 @@
 //  Unit conversions
 //
 #define RPM_TO_COMM_TIME(x) (1000000UL * 10U / (x) * TICKS_PER_US)
-//#define PWR_PCT_TO_VAL(x)   ((x) * SDM_TOP / 100U)
-#define US_TO_TICKS(x)      ((x) * TICKS_PER_US)
+#define US_TO_TICKS(x)      ((x) * (uint8_t)TICKS_PER_US)
 
 ////////////////////////////////////////////////////////////////////////////////
 ///
